requirements.txt
kafka-python==2.2.15
python-kasa==0.10.2
tapo==0.8.6
